Manual upload of all CURRENT files + Snake working
This are pretty much all the Files needed for this game, I am still manually uploading them since I want to focus on the game and then try to fix this repository. Preferably change the init location to Resources and try to add the libraries, which contain thousands of files to a .gitignore.